+++ CromWood/Controllers/RolePermissionController.cs
@@ -1,15 +1,20 @@
-﻿using CromWood.Business.Models;
+﻿using CromWood.Business.Constants;
+using CromWood.Business.Models;
 using CromWood.Business.Services.Interface;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 
 namespace CromWood.Controllers
 {
+    [Authorize]
     public class RolePermissionController : Controller
     {
         private readonly IRolePermissionService _rolePermissionService;
-        public RolePermissionController(IRolePermissionService rolePermissionService)
+        private readonly IAuthService _authService;
+        public RolePermissionController(IRolePermissionService rolePermissionService, IAuthService authService)
         {
             _rolePermissionService = rolePermissionService;
+            _authService = authService;
         }
 
         /// <summary>
@@ -17,6 +22,11 @@ namespace CromWood.Controllers
         /// </summary>
         public async Task<IActionResult> Index()
         {
+            var havePermission = await _authService.CheckPermission(PermissionKeyConstant.RoleManagement, PermissionConstant.ViewAll);
+            if (!havePermission)
+            {
+                return RedirectToAction("NotAuthorized", "Auth");
+            }
             var roles = await _rolePermissionService.GetRolesAsync();
             return View(roles.Data);
         }
@@ -27,12 +37,19 @@ namespace CromWood.Controllers
         [HttpGet]
         public async Task<IActionResult> AddModifyRole(Guid Id)
         {
+            var havePermission = await _authService.CheckPermission(PermissionKeyConstant.RoleManagement, PermissionConstant.CanWrite);
+            if (!havePermission)
+            {
+                return RedirectToAction("NotAuthorized", "Auth");
+            }
+
             var role = new RoleModel();
+            var permissions = await _rolePermissionService.GetPermissionsAsync();
+
             if (Id == Guid.Empty)
             {
-                var permissions = await _rolePermissionService.GetPermissionsAsync();
                 var rolePermissions = new List<RolePermissionModel>();
-                foreach (var permission in permissions.Data.Where(x => x.PermissionKey != "all_permission"))
+                foreach (var permission in permissions.Data)
                 {
                     rolePermissions.Add(new RolePermissionModel()
                     {
@@ -44,6 +61,16 @@ namespace CromWood.Controllers
             else
             {
                 var result = await _rolePermissionService.GetRoleByIdAsync(Id);
+                foreach (var permission in permissions.Data)
+                {
+                    if (!result.Data.RolePermission.Where(x => x.Permission.PermissionKey == permission.PermissionKey).Any())
+                    {
+                        result.Data.RolePermission.Add(new RolePermissionModel()
+                        {
+                            Permission = permission,
+                        });
+                    }
+                }
                 role = result.Data;
             }
 
@@ -56,6 +83,12 @@ namespace CromWood.Controllers
         [HttpPost]
         public async Task<IActionResult> AddModifyRole([FromForm] RoleModel role)
         {
+            var havePermission = await _authService.CheckPermission(PermissionKeyConstant.RoleManagement, PermissionConstant.CanWrite);
+            if (!havePermission)
+            {
+                return RedirectToAction("NotAuthorized", "Auth");
+            }
+
             if (role.Id == Guid.Empty)
             {
                 await _rolePermissionService.AddRoleAsync(role);
@@ -71,8 +104,13 @@ namespace CromWood.Controllers
         /// GET: This method will show modal to confirm role deletion
         /// </summary>
         [HttpGet]
-        public IActionResult DeleteRoleModal(Guid Id)
+        public async Task<IActionResult> DeleteRoleModal(Guid Id)
         {
+            var havePermission = await _authService.CheckPermission(PermissionKeyConstant.RoleManagement, PermissionConstant.CanDelete);
+            if (!havePermission)
+            {
+                return RedirectToAction("NotAuthorized", "Auth");
+            }
             return PartialView("DeleteRole", Id);
         }
 
@@ -81,6 +119,11 @@ namespace CromWood.Controllers
         /// </summary>
         public async Task<IActionResult> DeleteRole(Guid Id)
         {
+            var havePermission = await _authService.CheckPermission(PermissionKeyConstant.RoleManagement, PermissionConstant.CanWrite);
+            if (!havePermission)
+            {
+                return RedirectToAction("NotAuthorized", "Auth");
+            }
             await _rolePermissionService.DeleteRoleAync(Id);
             return RedirectToAction("Index");
         }
